fnalign could not see lq/sq, so whole prologues compared equal
Capstone's MIPS64 mode has no `lq`/`sq`, the PS2's 128-bit load and store
(major opcodes 0x1E and 0x1F).  It returns nothing for those words, and both
`fnalign` and `fndiff` turned an empty decode into the literal `"??"`.

Every `"??"` compares equal to every other `"??"`.  MWCC spills callee-saved
registers with `sq`, so on any function with four or more saved registers the
prologue and epilogue were INVISIBLE to the aligner - a body that saved the
wrong registers, or a different number of them, cost nothing.

`func_001ee610` is how it surfaced.  Its prologue diff read

    retail  addiu $sp, $sp, -0x4e0        object  addiu $sp, $sp, -0x4c0
    retail  sd $ra, 0x70($sp)             object  sd $ra, 0x50($sp)
    retail  ??
    retail  subu.qb $zero, $sp, $s4

and those last two are `sq $s5, 0x60($sp)` and `sq $s4, 0x50($sp)`.  The
32-byte frame difference and the missing register saves are the whole story
of that floor, and the tool was printing them as noise.

tools/eedis.py decodes both opcodes and is tried BEFORE capstone, so a future
capstone that decodes them as some MIPS DSP instruction cannot silently win.
Ten tests cover the register and displacement fields, negative offsets, short
input, that ordinary instructions are still left to capstone, and - the point
of the exercise - that two different `sq` rows no longer compare equal and
that a missing save now costs an edit.

Checked for stale numbers: func_00304580 (2279), func_00235520 (1636) and
func_00131a00 (1833) are unchanged, because both sides of those floors had
the same save pattern and the `"??"` rows happened to line up.  The fix is a
pure gain in sensitivity, not a renumbering.  func_001ee610 now reports 380
edits against a prologue it can actually read.

Also this commit:

func_00443010 (strstr) 26 -> 25 instrs, 25 -> 21 edits, +18.2% -> +13.6%.
Its note claimed "all 8 singles + all 28 pragma pairs" had been swept and
that "the entry movn is the wall".  The sweep was right, the conclusion was
not: two instructions came out of the BODY with no pragma involved.  Retail
fills the `beqz` delay slot with `addu $t6,$v0,$t4`, the address of
`result[counter]` computed before the test that does not need it - hoisting
`p2 = result + counter;` above the test reproduces that.  And retail reaches
its epilogue by FALLING THROUGH `daddu $2,$0,$0`, so `result = NULL;` has to
be the last statement before a shared label, not an early break.  Still
outside the band; the back-edge `bnel` remains.

func_001ee610 diagnosed and recorded, not yet fixed: deficit 15, and all
three retail-only runs (37, 29, 23) are LONGER than the deficit, so every one
is a CROSS.  The body needs re-syncing, not new statements - starting with
the frame size and the extra live pointer.

Image 3d1d3d2b, SLUS 4eeec036, 656 tests, lint 0 errors / 286 warnings.

+++ src/Event/Fcl/y_fclCombineDraw.c
@@ -405,8 +405,8 @@ void func_003147e0(u8 *arg0, s8 arg1, FclVec2 arg2, s16 arg3, s32 arg4, s32 arg5
     f32 f20;
     f32 f21;
     s32 i;
-    s64 li;
-    s64 lia;
+    s32 li;
+    s32 lia;
     t = *(u8 **)(arg0 + 0x38);
     v19 = (s8)arg1;
     c0 = (s16)(v19 * 5 + 0x66);
@@ -465,7 +465,7 @@ void func_003147e0(u8 *arg0, s8 arg1, FclVec2 arg2, s16 arg3, s32 arg4, s32 arg5
         f20 = 6.0f + arg2.x;
         i = 0;
         while ((s16)i < 2) {
-            li = (s64)(s16)i;
+            li = (s16)i;
             lia = (s16)(li + 0x174);
             func_002b2970((u8 *)&sp118, f20 + (f32)(li * 0x67), f21);
             func_002b2a60(&c150, 0xCC, 0xFF, 0xFF, 0xFF);
@@ -1286,7 +1286,7 @@ void func_00317410(u8 *arg0, s8 arg1) {
 // FUN_00317900 NONMATCHING
 #ifdef NON_MATCHING
 #pragma opt_propagation off
-void func_00317900(u8 *arg0, s64 arg1, s64 arg2, s64 arg3, s64 arg4, s64 arg5, s64 arg6) {
+void func_00317900(u8 *arg0, s64 arg1, s64 arg2, s64 arg3, s64 arg4, s32 arg5, s64 arg6) {
     u8 c23C[4];
     u8 c238[4];
     u8 c234[4];
@@ -1331,7 +1331,7 @@ void func_00317900(u8 *arg0, s64 arg1, s64 arg2, s64 arg3, s64 arg4, s64 arg5, s
     s64 sp108;
     FclPackedPosition sp100;
     s64 spF8;
-    s64 spD0;
+    s32 spD0;
     s32 v18;
     s32 t22;
     s32 t16;
@@ -2232,7 +2232,7 @@ void func_0031ac10(u8 *arg0, s64 arg1, s8 arg2, u16 arg3, u16 arg4, s32 arg5, s3
     s16 id6;
     s16 lateId;
     s16 spE0;
-    s64 t16;
+    s32 t16;
     s64 t7;
     s8 v3;
     s8 v30;
@@ -2249,7 +2249,7 @@ void func_0031ac10(u8 *arg0, s64 arg1, s8 arg2, u16 arg3, u16 arg4, s32 arg5, s3
     f22 = by + (f32)(t16 * 0x17);
     t7 = *(s8 *)&arg7;
     if (t7 == 1) {
-        v3 = *(s8 *)(func_002e4870(0) + (s8)arg2 * 0xC + (s32)t16 + 0x14);
+        v3 = *(s8 *)(func_002e4870(0) + (s8)arg2 * 0xC + t16 + 0x14);
     }
     if (*(s8 *)(t + 0x128) != t16 && *(s8 *)(t + 0x129) != t16 && v3 > 0) {
         v30 = v3;
